refactor(probe): migrate dhcp planning into per-protocol plugin
